Shuffle some things around
