+++ src/nexgen/command_line/nxs_phil.py
@@ -176,7 +176,7 @@ beamline_scope = freephil.parse(
       type = Synchrotron X-ray Source
         .type = str
         .help = "Facility type"
-      beamline_name = "I19-2"
+      beamline_name = I19-2
         .type = str
         .help = "Beamline name"
     }
